+++ CHANGES.rst
@@ -5,8 +5,9 @@ Changelog of lizard-wbconfiguration
 0.1.12 (unreleased)
 -------------------
 
-- Added functionality to validate wbconfigurations for separate areas.
+- Added functionality to import wbconfigurations using validation configuration.
 
+- Created task to validate wbconfigurations.
 
 0.1.11 (2012-03-26)
 -------------------